Add event sharing and hash change handler
Also supports being given GPS coords and placing a marker
